chore(front): move BackendService interface
