--- src/lib/backend-response.ts
@@ -1,13 +1,0 @@
-export type BackendError = {
-  type: 'error';
-  message: string;
-};
-
-export type BackendWarning = {
-  type: 'warning';
-  message: string;
-};
-
-export type BackendResponse<T> =
-  | Promise<{ data: T; error?: never; warning?: BackendWarning[] }>
-  | Promise<{ data?: never; error: BackendError[] }>;
